+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 edbee.lib:
 
+- add #121, Insert line before and insert line after commands
 - add #108, #111, Added a DynamicTextRangeSet, a change aware rangeset, that automatically gets adjusted when the document changes.
 - add #107, Implemented scoped/unscoped environment variable support
 - add #33, Toggle comment line and comment block support using the TM_COMMENT_* environment variable structure.